backend/test/unit/modules/lc/assets/cancel_asset.spec.ts
import { CancelAsset } from '../../../../../src/app/modules/lc/assets/cancel_asset';

describe('CancelAsset', () => {
  let transactionAsset: CancelAsset;

	beforeEach(() => {
		transactionAsset = new CancelAsset();
	});

	describe('constructor', () => {
		it('should have valid id', () => {
			expect(transactionAsset.id).toEqual(4);
		});

		it('should have valid name', () => {
			expect(transactionAsset.name).toEqual('cancel');
		});

		it('should have valid schema', () => {
			expect(transactionAsset.schema).toMatchSnapshot();
		});
	});

	describe('validate', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('apply', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
